# api/openapi.yaml
openapi: 3.0.3
info:
  title: loghub API
  version: 0.1.0
  description: |
    Multi-tenant log management: ingest events from many sources, normalize them
    into one schema, search and count them, and raise alerts when they repeat.

    ## Shape of this API

    Ingest is permissive and search is strict. A sender may post whatever its
    vendor emits; unknown fields are preserved verbatim in `raw` and stay
    searchable. What comes back out of search is the normalized schema, with
    every field named as in the common event schema.

    ## Errors

    Every error is the same JSON envelope: a stable `code` to match on, a
    human `message` that may change, and `request_id` correlating with the
    server log. `422` additionally carries per-field failures in `errors[]`;
    no other status does.

    A `404` or `405` from an unrouted path is generated by the Go HTTP mux and
    is plain text rather than this envelope.

    ## Authentication

    People sign in with `POST /api/v1/auth/login` and send the token it
    returns as `Authorization: Bearer <token>`. A token lasts 12 hours and
    cannot be refreshed; sign in again when it expires.

    Senders that aren't people, such as the collector and the sample scripts,
    use the ingest key instead: `INGEST_TOKEN` from the appliance's `.env`,
    sent the same way. It may send events for any tenant and do nothing else.

    There are two roles. An `admin` reads every tenant, may ingest for any of
...

...

...
servers:
...
paths:
...
  /api/v1/events/top:
    get:
      operationId: topValues
      summary: Most frequent values
      description: |
        Counts matching events by the value of one field, and returns the most
        frequent values. Events with no value for the field aren't counted.
        Values with the same count are in byte order.

        Unusable parameters and timeouts are answered as search answers them.
      tags: [Dashboard]
      security:
        - sessionToken: []
      parameters:
        - $ref: '#/components/parameters/TopField'
        - $ref: '#/components/parameters/TopLimit'
        - $ref: '#/components/parameters/TenantFilter'
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - $ref: '#/components/parameters/SourceFilter'
        - $ref: '#/components/parameters/EventTypeFilter'
        - $ref: '#/components/parameters/ActionFilter'
        - $ref: '#/components/parameters/SeverityMin'
        - $ref: '#/components/parameters/SeverityMax'
        - $ref: '#/components/parameters/SrcIpFilter'
        - $ref: '#/components/parameters/UserFilter'
        - $ref: '#/components/parameters/HostFilter'
        - $ref: '#/components/parameters/TagFilter'
        - $ref: '#/components/parameters/Query'
      responses:
        '200':
          description: The most frequent values, most frequent first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopValueList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/events/timeline:
    get:
      operationId: timeline
      summary: Events over time
      description: |
        Counts matching events in buckets of equal width, for a chart of the
        window.

        Without `interval`, the shortest interval that divides the window
        into at most 200 parts is used: one minute for an hour, 15 minutes for
        a day, an hour for a week. A window more than 1440 intervals long is
        refused with `invalid_parameter`, so a day can be counted by the
        minute but not two.

        Other unusable parameters and timeouts are answered as search answers
        them.
      tags: [Dashboard]
      security:
        - sessionToken: []
      parameters:
        - $ref: '#/components/parameters/TimelineInterval'
        - $ref: '#/components/parameters/TenantFilter'
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - $ref: '#/components/parameters/SourceFilter'
        - $ref: '#/components/parameters/EventTypeFilter'
        - $ref: '#/components/parameters/ActionFilter'
        - $ref: '#/components/parameters/SeverityMin'
        - $ref: '#/components/parameters/SeverityMax'
        - $ref: '#/components/parameters/SrcIpFilter'
        - $ref: '#/components/parameters/UserFilter'
        - $ref: '#/components/parameters/HostFilter'
        - $ref: '#/components/parameters/TagFilter'
        - $ref: '#/components/parameters/Query'
      responses:
        '200':
          description: Every bucket in the window, oldest first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Timeline'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /api/v1/alerts:
    get:
      operationId: listAlerts
      summary: List alerts
      description: The newest alerts the caller may read, newest first.
      tags: [Alerts]
      security:
        - sessionToken: []
      parameters:
        - $ref: '#/components/parameters/TenantFilter'
        - $ref: '#/components/parameters/AlertLimit'
      responses:
        '200':
          description: The alerts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
components:
  parameters:
    TenantFilter:
      name: tenant
      in: query
      required: false
      description: |
        Narrows results to one tenant the caller may read. An admin may read
        every tenant, and a viewer only their own. Naming a tenant outside
        that set is 403 `tenant_not_permitted`, not an empty page; a tenant
        within it that doesn't exist returns nothing.

        Omitting it is always safe and is the common case — it returns
        everything the caller may see.
      schema:
        type: string
        pattern: '^[A-Za-z0-9_-]{1,64}$'
      example: demoA
    From:
      name: from
      in: query
      required: false
      description: Only events at or after this time. Defaults to 24 hours ago.
      schema:
        type: string
        format: date-time
      example: '2025-08-20T00:00:00Z'
    To:
      name: to
      in: query
      required: false
      description: |
        Only events strictly before this time. Defaults to an hour from now,
        the furthest ahead an event time is accepted, so an event from a
        sender whose clock runs fast still shows up.
      schema:
        type: string
        format: date-time
    SourceFilter:
      name: source
      in: query
      required: false
      description: Repeat to match any of several sources.
      style: form
      explode: true
      schema:
        type: array
        items:
          $ref: '#/components/schemas/Source'
    EventTypeFilter:
      name: event_type
      in: query
      required: false
      schema:
        type: string
    ActionFilter:
      name: action
      in: query
      required: false
      schema:
        type: string
      example: deny
    SeverityMin:
      name: severity_min
      in: query
      required: false
      description: Only events with severity at or above this. Excludes events with no severity.
      schema:
        type: integer
        minimum: 0
        maximum: 10
    SeverityMax:
      name: severity_max
      in: query
      required: false
      description: Only events with severity at or below this. Excludes events with no severity.
      schema:
        type: integer
        minimum: 0
        maximum: 10
    SrcIpFilter:
      name: src_ip
      in: query
      required: false
      description: Exact source address match.
      schema:
        type: string
      example: 203.0.113.77
    UserFilter:
      name: user
      in: query
      required: false
      description: Exact match on the user associated with the event.
      schema:
        type: string
    HostFilter:
      name: host
      in: query
      required: false
      schema:
        type: string
    TagFilter:
      name: tag
      in: query
      required: false
      description: Repeat to require several tags; an event must carry all of them.
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
    Query:
      name: q
      in: query
      required: false
      description: |
        Free-text match against the original payload: any string or number
        value in `raw` that contains this text, ignoring case. Keys don't
        match. This is the slowest filter available, so narrow the time range
        with it.
      schema:
        type: string
...
    TopField:
      name: field
      in: query
      required: true
      description: The field to count events by.
      schema:
        $ref: '#/components/schemas/TopField'
    TopLimit:
      name: limit
      in: query
      required: false
      description: Maximum values to return. Defaults to 10.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
    TimelineInterval:
      name: interval
      in: query
      required: false
      description: |
        The width of each bucket. Defaults to the shortest that divides the
        window into at most 200 parts.
      schema:
        $ref: '#/components/schemas/TimelineInterval'
    AlertLimit:
      name: limit
      in: query
      required: false
      description: Maximum alerts to return. Defaults to 50.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 1000
  responses:
    BadRequest:
      description: The request could not be parsed or a parameter was unusable.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: invalid_parameter
            message: limit must be an integer between 1 and 1000
            request_id: 01J9Z3K2QW8XM4P7
    Unauthorized:
      description: |
        No credential was sent (`authentication_required`), or the one sent is
        malformed, expired or not recognised (`invalid_token`).
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: invalid_token
            message: the token is invalid or has expired; sign in again
            request_id: 01J9Z3K2QW8XM4P7
    Forbidden:
      description: |
        The caller is signed in but may not do this (`permission_denied`), or
        named a tenant it may not read (`tenant_not_permitted`).
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: tenant_not_permitted
            message: you may not access tenant "demoB"
            request_id: 01J9Z3K2QW8XM4P7
...
    InternalError:
      description: |
        Something failed on the server. The detail is in the logs; quote
        `request_id` when reporting it.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: internal_error
            message: an unexpected error occurred
            request_id: 01J9Z3K2QW8XM4P7
    ServiceUnavailable:
      description: |
        A dependency, normally the database, is unreachable, or a search or
        count ran past its time limit (`search_timeout`). Retrying later may
        succeed.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: database_unavailable
            message: cannot reach the database
            request_id: 01J9Z3K2QW8XM4P7
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The token from `POST /api/v1/auth/login`. Valid for 12 hours.
    ingestKey:
      type: http
      scheme: bearer
      description: |
        `INGEST_TOKEN` from the appliance's `.env`. It may send events for any
        tenant and nothing else.
  headers:
    XRequestId:
      description: Correlates this response with the server log line.
      schema:
        type: string
  schemas:
...
    Source:
      type: string
      description: |
        Source category. Closed set: the database rejects anything else, so an
        unrecognised source on ingest is a rejected record rather than a new
        category.
      enum:
        - firewall
        - network
        - crowdstrike
        - aws
        - m365
        - ad
        - api
...
    TopField:
      type: string
      description: |
        An event field to count by: `src_ip`, `dst_ip`, `user`, `host` or
        `event_type`.
      enum: [src_ip, dst_ip, user, host, event_type]
      example: src_ip
    TopValueList:
      type: object
      additionalProperties: false
      required: [field, from, to, items]
      properties:
        field:
          $ref: '#/components/schemas/TopField'
        from:
          type: string
          format: date-time
          description: Start of the window counted, inclusive.
        to:
          type: string
          format: date-time
          description: End of the window counted, exclusive.
        items:
          type: array
          description: The most frequent values, most frequent first.
          items:
            $ref: '#/components/schemas/TopValue'
    TopValue:
      type: object
      additionalProperties: false
      required: [value, count]
      properties:
        value:
          type: string
          description: A value of the field, as search returns it.
          example: 203.0.113.77
        count:
          type: integer
          format: int64
          description: How many matching events had this value.
          example: 42
    TimelineInterval:
      type: string
      description: |
        A bucket width: minutes (`m`), hours (`h`) or a day (`d`). Buckets
        start on whole multiples of the width in UTC, so days start at
        midnight UTC.
      enum: [1m, 5m, 15m, 30m, 1h, 3h, 6h, 12h, 1d]
      example: 15m
    Timeline:
      type: object
      additionalProperties: false
      required: [from, to, interval, buckets]
      properties:
        from:
          type: string
          format: date-time
          description: Start of the window counted, inclusive.
        to:
          type: string
          format: date-time
          description: End of the window counted, exclusive.
        interval:
          $ref: '#/components/schemas/TimelineInterval'
        buckets:
          type: array
          description: |
            Every bucket that overlaps the window, oldest first, empty ones
            included. The first may start before `from` and the last may end
            after `to`, but only events within the window are counted. So
            there may be one more bucket than the window has intervals.
          items:
            $ref: '#/components/schemas/TimelineBucket'
    TimelineBucket:
      type: object
      additionalProperties: false
      required: [start, count]
      properties:
        start:
          type: string
          format: date-time
          description: |
            Start of the bucket, inclusive. It ends where the next one starts.
        count:
          type: integer
          format: int64
          description: How many matching events fell in the bucket.
          example: 17
    AlertGroupBy:
      type: string
      description: |
        The event field a rule counts by: `src_ip`, `dst_ip`, `user` or
        `host`. Events without a value for it are not counted.
      enum: [src_ip, dst_ip, user, host]
...
    Alert:
      type: object
      additionalProperties: false
      required: [id, rule_id, rule_name, tenant, group_by, group_key, count, window_start, window_end, created_at]
      description: |
        One firing of a rule: `count` matching events whose `group_by` field
        was `group_key` fell between `window_start` and `window_end`. This is
        also the body a rule's webhook receives.
      properties:
        id:
          type: string
          example: '301'
        rule_id:
          type: string
          example: '12'
        rule_name:
          type: string
          example: Repeated failed logins from one address
        tenant:
          type: string
          example: demoA
        group_by:
          $ref: '#/components/schemas/AlertGroupBy'
        group_key:
          type: string
          description: The value the matching events shared, such as an address.
          example: 203.0.113.77
        count:
          type: integer
          format: int64
          description: How many matching events the window held.
          example: 5
        window_start:
          type: string
          format: date-time
          description: Start of the window, inclusive.
        window_end:
          type: string
          format: date-time
          description: End of the window, exclusive.
        created_at:
          type: string
          format: date-time
          description: When the alert was raised.
    AlertList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Alert'
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            Stable, machine-readable. Match on this; `message` is for people
            and may be reworded at any time.
          example: invalid_request
        message:
          type: string
        request_id:
          type: string
          description: Also returned as the `X-Request-Id` header.
        errors:
          type: array
          description: Per-field failures. Present only on 422.
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      additionalProperties: false
      required: [field, code, message]
      properties:
        field:
          type: string
          example: severity
        code:
          type: string
          example: out_of_range
        message:
          type: string